Dog Walk
fetch all Dog entities with names of “Ruby” from Core Data. If the
fetch request comes back with zero results, this probably means it’s
the user’s first time opening the app. If this is the case, you insert
a new dog, name it “Ruby”, and set it as the currently selected dog.
If the fetch request came back with results, you set the first entity
(there should only be one) as the currently selected dog.