Dog Walk (Summary)
setup and the underlying pieces that make up Core Data. This was
intentional! Core Data has a reputation for its steep learning curve.
This is partly because of all the setup it requires just to get
started: the stack, the data model, the managed object subclasses, et
cetera. In addition, you got some firsthand experience with
relationships and deletion.

func tableView(tableView: UITableView, didSelectRowAtIndexPath
indexPath: NSIndexPath) {
//        [tableView deselectRowAtIndexPath:indexPath animated:YES];
        tableView.deselectRowAtIndexPath(indexPath, animated: true)

    }