Bubble Tea(Asynchronous fetching)
A managed object context can have one of three concurrency types. If
you don’t specify a concurrency type, it defaults to
.ConfinementConcurrencyType. The problem is that this concurrency type
is an old pattern that is all but deprecated.
As such, asynchronous fetching doesn’t work with the confinement
concurrency type, so you have to be explicit and set your concurrency
type to .MainQueueConcurrencyType.