World Cup
Setting up a fetched results controller
1. The fetched results controller handles the coordination between Core
Data and your table view, but it still needs you to provide an
NSFetchRequest. Remember that the NSFetchRequest class is highly
customizable. It can take sort descriptors, predicates, etc.
In this example, you simply initialize a fetch request with an entity
description, because you want to fetch all Team objects.
2. The initializer method for a fetched results controller takes four
parameters: first is the fetch request you just created.
The second parameter is an instance of NSManagedObjectContext. Like
NSFetchRequest, the fetched results controller class needs a managed
object context to execute the fetch. It can’t actually fetch anything
by itself.
The other two parameters are optional: sectionNameKeyPath and cacheName.
3. You execute the fetch request. If there’s an error, you log the
error to the console.