UnCloudNotes (Enabling lightweight migrations)
Enabling lightweight migrations

To enable lightweight migrations, you need to set two flags on
initialization.

The NSMigratePersistentStoresAutomaticallyOption is what tells Core
Data (the NSPersistentStoreCoordinator, actually) to start a migration
if the persistent store model isn’t compatible with the current data
model. Core Data will handle all the details, from finding the source
model to creating the destination store file, and all the steps in
between.
The NSInferMappingModelAutomaticallyOption is the other half of what
makes a lightweight migration possible. Every migration requires a
mapping model. Here’s an analogy: If you’re traveling from a known
place on Earth to somewhere unknown, you’ll want a map to tell you
where to go. The mapping model is that guide.
It just so happens that Core Data can infer a mapping model in many
cases. That is, Core Data can automatically look at the differences in
two data models and create a mapping model between them. For entities
and attributes that are identical between model versions, this is a
straightforward data pass through mapping. For other changes, just
follow a few simple rules for Core Data to create a mapping
model. In the new model, changes must fit an obvious migration pattern,
such as:
1. Deleting entities, attributes or relationships;
2. Renaming entities, attributes or relationships using the
renamingIdentifier;
3. Adding a new, optional attribute;
4. Adding a new, required attribute with a default value;
5. Changing an optional attribute to non-optional and specifying a
default value;
6. Changing a non-optional attribute to optional;
7. Changing the entity hierarchy;
8. Adding a new parent entity and moving attributes up or down the
hierarchy;
9. Changing a relationship from to-one to to-many;
10. Changing a relationship from non-ordered to-many to ordered to-many
(and vice versa).