UnCloudNotes (Custom migration policies)
AttachmentToImageAttachmentMigrationPolicyV3toV4.swift

1. First, you create an instance of the new destination object. The
migration manager has two Core Data stacks—one to read from the source
and one to write to the destination—so you need to be sure to use the
destination context here.
2. Even though this is a custom manual migration, most of the attribute
migrations should be performed using the expressions you defined in the
mapping model. To do this, you loop through each attribute mapping from
the entityMapping parameter, apply the value expression to the source
instance and set the result to the new destination object.
3. Next, you try to get an instance of the image. If it exists, then
you grab its width and height to populate the data in the new object.
4. For the caption, you simply grab the note’s body text and take the
first 80 characters.
5. The migration manager needs to know the connection between the
source object, the newly created destination object and the mapping.
Failing to call this method at the end of a custom migration will
result in missing data in the destination store.
6. Finally, the return value tells the migration manager that
everything went well.